+++ transports/websocket/src/framed.rs
@@ -21,11 +21,11 @@
 use crate::{error::Error, quicksink, tls};
 use either::Either;
 use futures::{future::BoxFuture, prelude::*, ready, stream::BoxStream};
-use futures_rustls::{client, rustls, server};
+use futures_rustls::rustls::pki_types::ServerName;
+use futures_rustls::{client, server};
 use libp2p_core::{
-    connection::Endpoint,
     multiaddr::{Multiaddr, Protocol},
-    transport::{ListenerId, TransportError, TransportEvent},
+    transport::{DialOpts, ListenerId, TransportError, TransportEvent},
     Transport,
 };
 use parking_lot::Mutex;
@@ -34,6 +34,7 @@ use soketto::{
     handshake,
 };
 use std::borrow::Cow;
+use std::net::IpAddr;
 use std::{collections::HashMap, ops::DerefMut, sync::Arc};
 use std::{fmt, io, mem, pin::Pin, task::Context, task::Poll};
 use url::Url;
@@ -141,19 +142,12 @@ where
         self.transport.lock().remove_listener(id)
     }
 
-    fn dial(&mut self, addr: Multiaddr) -> Result<Self::Dial, TransportError<Self::Error>> {
-        self.do_dial(addr, Endpoint::Dialer)
-    }
-
-    fn dial_as_listener(
+    fn dial(
         &mut self,
         addr: Multiaddr,
+        dial_opts: DialOpts,
     ) -> Result<Self::Dial, TransportError<Self::Error>> {
-        self.do_dial(addr, Endpoint::Listener)
-    }
-
-    fn address_translation(&self, server: &Multiaddr, observed: &Multiaddr) -> Option<Multiaddr> {
-        self.transport.lock().address_translation(server, observed)
+        self.do_dial(addr, dial_opts)
     }
 
     fn poll(
@@ -249,7 +243,7 @@ where
     fn do_dial(
         &mut self,
         addr: Multiaddr,
-        role_override: Endpoint,
+        dial_opts: DialOpts,
     ) -> Result<<Self as Transport>::Dial, TransportError<<Self as Transport>::Error>> {
         let mut addr = match parse_ws_dial_addr(addr) {
             Ok(addr) => addr,
@@ -268,8 +262,7 @@ where
 
         let future = async move {
             loop {
-                match Self::dial_once(transport.clone(), addr, tls_config.clone(), role_override)
-                    .await
+                match Self::dial_once(transport.clone(), addr, tls_config.clone(), dial_opts).await
                 {
                     Ok(Either::Left(redirect)) => {
                         if remaining_redirects == 0 {
@@ -293,33 +286,29 @@ where
         transport: Arc<Mutex<T>>,
         addr: WsAddress,
         tls_config: tls::Config,
-        role_override: Endpoint,
+        dial_opts: DialOpts,
     ) -> Result<Either<String, Connection<T::Output>>, Error<T::Error>> {
         tracing::trace!(address=?addr, "Dialing websocket address");
 
-        let dial = match role_override {
-            Endpoint::Dialer => transport.lock().dial(addr.tcp_addr),
-            Endpoint::Listener => transport.lock().dial_as_listener(addr.tcp_addr),
-        }
-        .map_err(|e| match e {
-            TransportError::MultiaddrNotSupported(a) => Error::InvalidMultiaddr(a),
-            TransportError::Other(e) => Error::Transport(e),
-        })?;
+        let dial = transport
+            .lock()
+            .dial(addr.tcp_addr, dial_opts)
+            .map_err(|e| match e {
+                TransportError::MultiaddrNotSupported(a) => Error::InvalidMultiaddr(a),
+                TransportError::Other(e) => Error::Transport(e),
+            })?;
 
         let stream = dial.map_err(Error::Transport).await?;
         tracing::trace!(port=%addr.host_port, "TCP connection established");
 
         let stream = if addr.use_tls {
             // begin TLS session
-            let dns_name = addr
-                .dns_name
-                .expect("for use_tls we have checked that dns_name is some");
-            tracing::trace!(?dns_name, "Starting TLS handshake");
+            tracing::trace!(?addr.server_name, "Starting TLS handshake");
             let stream = tls_config
                 .client
-                .connect(dns_name.clone(), stream)
+                .connect(addr.server_name.clone(), stream)
                 .map_err(|e| {
-                    tracing::debug!(?dns_name, "TLS handshake failed: {}", e);
+                    tracing::debug!(?addr.server_name, "TLS handshake failed: {}", e);
                     Error::Tls(tls::Error::from(e))
                 })
                 .await?;
@@ -481,7 +470,7 @@ impl<'a> WsListenProto<'a> {
 struct WsAddress {
     host_port: String,
     path: String,
-    dns_name: Option<rustls::pki_types::ServerName<'static>>,
+    server_name: ServerName<'static>,
     use_tls: bool,
     tcp_addr: Multiaddr,
 }
@@ -498,19 +487,21 @@ fn parse_ws_dial_addr<T>(addr: Multiaddr) -> Result<WsAddress, Error<T>> {
     let mut protocols = addr.iter();
     let mut ip = protocols.next();
     let mut tcp = protocols.next();
-    let (host_port, dns_name) = loop {
+    let (host_port, server_name) = loop {
         match (ip, tcp) {
             (Some(Protocol::Ip4(ip)), Some(Protocol::Tcp(port))) => {
-                break (format!("{ip}:{port}"), None)
+                let server_name = ServerName::IpAddress(IpAddr::V4(ip).into());
+                break (format!("{ip}:{port}"), server_name);
             }
             (Some(Protocol::Ip6(ip)), Some(Protocol::Tcp(port))) => {
-                break (format!("{ip}:{port}"), None)
+                let server_name = ServerName::IpAddress(IpAddr::V6(ip).into());
+                break (format!("[{ip}]:{port}"), server_name);
             }
             (Some(Protocol::Dns(h)), Some(Protocol::Tcp(port)))
             | (Some(Protocol::Dns4(h)), Some(Protocol::Tcp(port)))
             | (Some(Protocol::Dns6(h)), Some(Protocol::Tcp(port)))
             | (Some(Protocol::Dnsaddr(h)), Some(Protocol::Tcp(port))) => {
-                break (format!("{}:{}", &h, port), Some(tls::dns_name_ref(&h)?))
+                break (format!("{h}:{port}"), tls::dns_name_ref(&h)?)
             }
             (Some(_), Some(p)) => {
                 ip = Some(p);
@@ -529,26 +520,14 @@ fn parse_ws_dial_addr<T>(addr: Multiaddr) -> Result<WsAddress, Error<T>> {
         match protocols.pop() {
             p @ Some(Protocol::P2p(_)) => p2p = p,
             Some(Protocol::Ws(path)) => match protocols.pop() {
-                Some(Protocol::Tls) => {
-                    if dns_name.is_none() {
-                        tracing::debug!(address=%addr, "Missing DNS name in WSS address");
-                        return Err(Error::InvalidMultiaddr(addr));
-                    }
-                    break (true, path.into_owned());
-                }
+                Some(Protocol::Tls) => break (true, path.into_owned()),
                 Some(p) => {
                     protocols.push(p);
                     break (false, path.into_owned());
                 }
                 None => return Err(Error::InvalidMultiaddr(addr)),
             },
-            Some(Protocol::Wss(path)) => {
-                if dns_name.is_none() {
-                    tracing::debug!(address=%addr, "Missing DNS name in WSS address");
-                    return Err(Error::InvalidMultiaddr(addr));
-                }
-                break (true, path.into_owned());
-            }
+            Some(Protocol::Wss(path)) => break (true, path.into_owned()),
             _ => return Err(Error::InvalidMultiaddr(addr)),
         }
     };
@@ -562,7 +541,7 @@ fn parse_ws_dial_addr<T>(addr: Multiaddr) -> Result<WsAddress, Error<T>> {
 
     Ok(WsAddress {
         host_port,
-        dns_name,
+        server_name,
         path,
         use_tls,
         tcp_addr,
@@ -864,75 +843,126 @@ mod tests {
 
     #[test]
     fn dial_addr() {
-        let tcp_addr = "/dns4/example.com/tcp/2222".parse::<Multiaddr>().unwrap();
-        let p2p = Protocol::P2p(PeerId::random());
+        let peer_id = PeerId::random();
 
         // Check `/tls/ws`
-        let addr = tcp_addr
-            .clone()
-            .with(Protocol::Tls)
-            .with(Protocol::Ws("/".into()));
-        let info = parse_ws_dial_addr::<io::Error>(addr.clone()).unwrap();
+        let addr = "/dns4/example.com/tcp/2222/tls/ws"
+            .parse::<Multiaddr>()
+            .unwrap();
+        let info = parse_ws_dial_addr::<io::Error>(addr).unwrap();
         assert_eq!(info.host_port, "example.com:2222");
         assert_eq!(info.path, "/");
         assert!(info.use_tls);
-        assert_eq!(info.dns_name, "example.com".try_into().ok());
-        assert_eq!(&info.tcp_addr, &tcp_addr);
+        assert_eq!(info.server_name, "example.com".try_into().unwrap());
+        assert_eq!(info.tcp_addr, "/dns4/example.com/tcp/2222".parse().unwrap());
 
         // Check `/tls/ws` with `/p2p`
-        let addr = tcp_addr
-            .clone()
-            .with(Protocol::Tls)
-            .with(Protocol::Ws("/".into()))
-            .with(p2p.clone());
+        let addr = format!("/dns4/example.com/tcp/2222/tls/ws/p2p/{peer_id}")
+            .parse()
+            .unwrap();
         let info = parse_ws_dial_addr::<io::Error>(addr).unwrap();
         assert_eq!(info.host_port, "example.com:2222");
         assert_eq!(info.path, "/");
         assert!(info.use_tls);
-        assert_eq!(info.dns_name, "example.com".try_into().ok());
-        assert_eq!(info.tcp_addr, tcp_addr.clone().with(p2p.clone()));
+        assert_eq!(info.server_name, "example.com".try_into().unwrap());
+        assert_eq!(
+            info.tcp_addr,
+            format!("/dns4/example.com/tcp/2222/p2p/{peer_id}")
+                .parse()
+                .unwrap()
+        );
+
+        // Check `/tls/ws` with `/ip4`
+        let addr = "/ip4/127.0.0.1/tcp/2222/tls/ws"
+            .parse::<Multiaddr>()
+            .unwrap();
+        let info = parse_ws_dial_addr::<io::Error>(addr).unwrap();
+        assert_eq!(info.host_port, "127.0.0.1:2222");
+        assert_eq!(info.path, "/");
+        assert!(info.use_tls);
+        assert_eq!(info.server_name, "127.0.0.1".try_into().unwrap());
+        assert_eq!(info.tcp_addr, "/ip4/127.0.0.1/tcp/2222".parse().unwrap());
+
+        // Check `/tls/ws` with `/ip6`
+        let addr = "/ip6/::1/tcp/2222/tls/ws".parse::<Multiaddr>().unwrap();
+        let info = parse_ws_dial_addr::<io::Error>(addr).unwrap();
+        assert_eq!(info.host_port, "[::1]:2222");
+        assert_eq!(info.path, "/");
+        assert!(info.use_tls);
+        assert_eq!(info.server_name, "::1".try_into().unwrap());
+        assert_eq!(info.tcp_addr, "/ip6/::1/tcp/2222".parse().unwrap());
 
         // Check `/wss`
-        let addr = tcp_addr.clone().with(Protocol::Wss("/".into()));
+        let addr = "/dns4/example.com/tcp/2222/wss"
+            .parse::<Multiaddr>()
+            .unwrap();
         let info = parse_ws_dial_addr::<io::Error>(addr).unwrap();
         assert_eq!(info.host_port, "example.com:2222");
         assert_eq!(info.path, "/");
         assert!(info.use_tls);
-        assert_eq!(info.dns_name, "example.com".try_into().ok());
-        assert_eq!(&info.tcp_addr, &tcp_addr);
+        assert_eq!(info.server_name, "example.com".try_into().unwrap());
+        assert_eq!(info.tcp_addr, "/dns4/example.com/tcp/2222".parse().unwrap());
 
         // Check `/wss` with `/p2p`
-        let addr = tcp_addr
-            .clone()
-            .with(Protocol::Wss("/".into()))
-            .with(p2p.clone());
+        let addr = format!("/dns4/example.com/tcp/2222/wss/p2p/{peer_id}")
+            .parse()
+            .unwrap();
         let info = parse_ws_dial_addr::<io::Error>(addr).unwrap();
         assert_eq!(info.host_port, "example.com:2222");
         assert_eq!(info.path, "/");
         assert!(info.use_tls);
-        assert_eq!(info.dns_name, "example.com".try_into().ok());
-        assert_eq!(info.tcp_addr, tcp_addr.clone().with(p2p.clone()));
+        assert_eq!(info.server_name, "example.com".try_into().unwrap());
+        assert_eq!(
+            info.tcp_addr,
+            format!("/dns4/example.com/tcp/2222/p2p/{peer_id}")
+                .parse()
+                .unwrap()
+        );
+
+        // Check `/wss` with `/ip4`
+        let addr = "/ip4/127.0.0.1/tcp/2222/wss".parse::<Multiaddr>().unwrap();
+        let info = parse_ws_dial_addr::<io::Error>(addr).unwrap();
+        assert_eq!(info.host_port, "127.0.0.1:2222");
+        assert_eq!(info.path, "/");
+        assert!(info.use_tls);
+        assert_eq!(info.server_name, "127.0.0.1".try_into().unwrap());
+        assert_eq!(info.tcp_addr, "/ip4/127.0.0.1/tcp/2222".parse().unwrap());
+
+        // Check `/wss` with `/ip6`
+        let addr = "/ip6/::1/tcp/2222/wss".parse::<Multiaddr>().unwrap();
+        let info = parse_ws_dial_addr::<io::Error>(addr).unwrap();
+        assert_eq!(info.host_port, "[::1]:2222");
+        assert_eq!(info.path, "/");
+        assert!(info.use_tls);
+        assert_eq!(info.server_name, "::1".try_into().unwrap());
+        assert_eq!(info.tcp_addr, "/ip6/::1/tcp/2222".parse().unwrap());
 
         // Check `/ws`
-        let addr = tcp_addr.clone().with(Protocol::Ws("/".into()));
+        let addr = "/dns4/example.com/tcp/2222/ws"
+            .parse::<Multiaddr>()
+            .unwrap();
         let info = parse_ws_dial_addr::<io::Error>(addr).unwrap();
         assert_eq!(info.host_port, "example.com:2222");
         assert_eq!(info.path, "/");
         assert!(!info.use_tls);
-        assert_eq!(info.dns_name, "example.com".try_into().ok());
-        assert_eq!(&info.tcp_addr, &tcp_addr);
+        assert_eq!(info.server_name, "example.com".try_into().unwrap());
+        assert_eq!(info.tcp_addr, "/dns4/example.com/tcp/2222".parse().unwrap());
 
         // Check `/ws` with `/p2p`
-        let addr = tcp_addr
-            .clone()
-            .with(Protocol::Ws("/".into()))
-            .with(p2p.clone());
+        let addr = format!("/dns4/example.com/tcp/2222/ws/p2p/{peer_id}")
+            .parse()
+            .unwrap();
         let info = parse_ws_dial_addr::<io::Error>(addr).unwrap();
         assert_eq!(info.host_port, "example.com:2222");
         assert_eq!(info.path, "/");
         assert!(!info.use_tls);
-        assert_eq!(info.dns_name, "example.com".try_into().ok());
-        assert_eq!(info.tcp_addr, tcp_addr.clone().with(p2p.clone()));
+        assert_eq!(info.server_name, "example.com".try_into().unwrap());
+        assert_eq!(
+            info.tcp_addr,
+            format!("/dns4/example.com/tcp/2222/p2p/{peer_id}")
+                .parse()
+                .unwrap()
+        );
 
         // Check `/ws` with `/ip4`
         let addr = "/ip4/127.0.0.1/tcp/2222/ws".parse::<Multiaddr>().unwrap();
@@ -940,26 +970,20 @@ mod tests {
         assert_eq!(info.host_port, "127.0.0.1:2222");
         assert_eq!(info.path, "/");
         assert!(!info.use_tls);
-        assert!(info.dns_name.is_none());
+        assert_eq!(info.server_name, "127.0.0.1".try_into().unwrap());
         assert_eq!(info.tcp_addr, "/ip4/127.0.0.1/tcp/2222".parse().unwrap());
 
-        // Check that `/tls/wss` is invalid
-        let addr = "/ip4/127.0.0.1/tcp/2222/tls/wss"
-            .parse::<Multiaddr>()
-            .unwrap();
-        parse_ws_dial_addr::<io::Error>(addr.clone()).unwrap_err();
-
-        // Check TLS addresses without domain
-        let addr = "/ip4/127.0.0.1/tcp/2222/tls/ws"
-            .parse::<Multiaddr>()
-            .unwrap();
-        parse_ws_dial_addr::<io::Error>(addr.clone()).unwrap_err();
-
-        let addr = "/ip4/127.0.0.1/tcp/2222/wss".parse::<Multiaddr>().unwrap();
-        parse_ws_dial_addr::<io::Error>(addr.clone()).unwrap_err();
+        // Check `/ws` with `/ip6`
+        let addr = "/ip6/::1/tcp/2222/ws".parse::<Multiaddr>().unwrap();
+        let info = parse_ws_dial_addr::<io::Error>(addr).unwrap();
+        assert_eq!(info.host_port, "[::1]:2222");
+        assert_eq!(info.path, "/");
+        assert!(!info.use_tls);
+        assert_eq!(info.server_name, "::1".try_into().unwrap());
+        assert_eq!(info.tcp_addr, "/ip6/::1/tcp/2222".parse().unwrap());
 
         // Check non-ws address
-        let addr = tcp_addr.clone().with(Protocol::Tls).with(p2p.clone());
-        parse_ws_dial_addr::<io::Error>(addr.clone()).unwrap_err();
+        let addr = "/ip4/127.0.0.1/tcp/2222".parse::<Multiaddr>().unwrap();
+        parse_ws_dial_addr::<io::Error>(addr).unwrap_err();
     }
 }