+++ transports/uds/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.41.0
+
+- Implement refactored `Transport`.
+  See [PR 4568](https://github.com/libp2p/rust-libp2p/pull/4568)
+
 ## 0.40.0