+++ transports/websocket/CHANGELOG.md
@@ -1,5 +1,9 @@
-## 0.43.3
+## 0.44.0
 
+- Implement refactored `Transport`.
+  See [PR 4568](https://github.com/libp2p/rust-libp2p/pull/4568)
+- Allow wss connections on IP addresses.
+  See [PR 5525](https://github.com/libp2p/rust-libp2p/pull/5525).
 - Add support for `/tls/ws` and keep `/wss` backward compatible.
   See [PR 5523](https://github.com/libp2p/rust-libp2p/pull/5523).